+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The present file will list all changes made to the project; according to the
 - Add ajax browsers notifications in addition to emails
 - Plugins can now add new notifications types (xmpp, sms, telegram, ...) to be used along with standard notifications
 - Simcard component
+- Synchronization field for LDAP
 
 ### Changed
 - Display knowledge base category items in tickets using a popup instead of a